Propagate to FRANKLIN:

Change 76304 by dkarp@dkarp-extra on 2007/12/06 15:16:39

	Handle MailItems representing items larger than 2GB.  Increase the 
	MAIL_ITEM.SIZE and REVISION.SIZE database columns to BIGINT.  
	Change MailItem.mData.size and MailItem.getSize() to return a long.
	Tweak a few things around the app to accommodate this, with the 
	only sketchy bit being that the methods returning item content in a 
	byte[] now must throw MailServiceException.MESSAGE_TOO_BIG on >2GB 
	items because Java doesn't support byte arrays larger than that.

	Add Folder.getItemCount() and try to call it instead of calling 
	Folder.getSize() when we want to know the number of non-subfolder 
	items in a folder.

	DB version number is now 50.  Please reset-the-world or run the 
	migrate20071206-WidenSizeColumns.pl migration script to get your 
	install's database schema in sync.

http://bugzilla.zimbra.com/show_bug.cgi?id=22362

Copied from Perforce
 Change: 76306

+++ src/java/com/zimbra/cs/dav/resource/CalendarObject.java
@@ -60,7 +60,7 @@ public class CalendarObject extends MailItemResource {
 		if (defInv != null)
 			setProperty(DavElements.P_DISPLAYNAME, defInv.getName());
 		setProperty(DavElements.P_GETCONTENTTYPE, Mime.CT_TEXT_CALENDAR);
-		setProperty(DavElements.P_GETCONTENTLENGTH, Integer.toString(calItem.getSize()));
+		setProperty(DavElements.P_GETCONTENTLENGTH, Long.toString(calItem.getSize()));
 		addProperty(CalDavProperty.getCalendarData(this));
 		mAccount = calItem.getAccount();
 	}
